+++ src/main/java/com/cemre/product_backend/repository/ProductRepository.java
@@ -4,6 +4,6 @@ import com.cemre.product_backend.model.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ProductRepository extends JpaRepository<Product, Long> {}
+public interface ProductRepository extends JpaRepository<Product, Integer> {}
